move fast inverse to main finite field code
