+++ crypto/benchmarks/inverse.unit-test.ts
@@ -1,18 +1,16 @@
-import { Fp } from '../finite_field.js';
+import { Fp, p, inverse } from '../finite_field.js';
 import { tic, toc } from '../../../examples/utils/tic-toc.node.js';
-import { createFastInverse } from './fast-inverse.js';
 
 const N = 10000;
 
 let fields: bigint[] = Array(N);
-const fastInverse = createFastInverse(Fp.modulus);
 
 bench('inverse', fillRandomFields, () => {
-  for (let i = 0; i < N; i++) fields[i] = Fp.inverse(fields[i])!;
+  for (let i = 0; i < N; i++) fields[i] = inverse(fields[i], p)!;
 });
 
 bench('fast inverse', fillRandomFields, () => {
-  for (let i = 0; i < N; i++) fields[i] = fastInverse(fields[i]);
+  for (let i = 0; i < N; i++) fields[i] = Fp.inverse(fields[i])!;
 });
 
 // helpers
